C11-159: extract theme/system/auth/app/snapshot/conversation/notification/markdown/feedback/settings/sidebar/session/tab/mailbox socket handlers
Mechanical relocation (DX-1..5, zero behavior change): move each domain's
processV2Command case block + its handler methods, verbatim, into per-domain
extension files under Sources/SocketHandlers/. Leaf handlers stay private in
their domain file; the v2DispatchExtracted router delegates by method prefix,
falling back to the identical method_not_found response. Files added to the
c11 target only. TerminalController.swift: 20351 -> 17959 LOC.
