+++ Sources/SocketHandlers/NotificationHandlers.swift
@@ -1,0 +1,149 @@
+import AppKit
+import Foundation
+
+// C11-159: per-domain socket handler unit extracted verbatim from
+// TerminalController.swift. Mechanical relocation, zero behavior change.
+extension TerminalController {
+    /// v2 dispatch slice for the `notification.*` domain(s).
+    /// Byte-identical routing and wire responses to the original processV2Command cases.
+    func v2DispatchNotification(_ method: String, id: Any?, params: [String: Any]) -> String {
+        switch method {
+        case "notification.create":
+            return v2Result(id: id, self.v2NotificationCreate(params: params))
+        case "notification.create_for_surface":
+            return v2Result(id: id, self.v2NotificationCreateForSurface(params: params))
+        case "notification.create_for_target":
+            return v2Result(id: id, self.v2NotificationCreateForTarget(params: params))
+        case "notification.list":
+            return v2Ok(id: id, result: self.v2NotificationList())
+        case "notification.clear":
+            return v2Result(id: id, self.v2NotificationClear())
+        default:
+            return v2Error(id: id, code: "method_not_found", message: "Unknown method")
+        }
+    }
+
+    private func v2NotificationCreate(params: [String: Any]) -> V2CallResult {
+        guard let tabManager = v2ResolveTabManager(params: params) else {
+            return .err(code: "unavailable", message: "TabManager not available", data: nil)
+        }
+
+        let title = (params["title"] as? String) ?? "Notification"
+        let subtitle = (params["subtitle"] as? String) ?? ""
+        let body = (params["body"] as? String) ?? ""
+
+        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
+        v2MainSync {
+            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
+                result = .err(code: "not_found", message: "Workspace not found", data: nil)
+                return
+            }
+            let surfaceId = ws.focusedPanelId
+            TerminalNotificationStore.shared.addNotification(
+                tabId: ws.id,
+                surfaceId: surfaceId,
+                title: title,
+                subtitle: subtitle,
+                body: body
+            )
+            result = .ok(["workspace_id": ws.id.uuidString, "surface_id": v2OrNull(surfaceId?.uuidString)])
+        }
+        return result
+    }
+
+    private func v2NotificationCreateForSurface(params: [String: Any]) -> V2CallResult {
+        guard let tabManager = v2ResolveTabManager(params: params) else {
+            return .err(code: "unavailable", message: "TabManager not available", data: nil)
+        }
+        guard let surfaceId = v2UUID(params, "surface_id") else {
+            return .err(code: "invalid_params", message: "Missing or invalid surface_id", data: nil)
+        }
+
+        let title = (params["title"] as? String) ?? "Notification"
+        let subtitle = (params["subtitle"] as? String) ?? ""
+        let body = (params["body"] as? String) ?? ""
+
+        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
+        v2MainSync {
+            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
+                result = .err(code: "not_found", message: "Workspace not found", data: nil)
+                return
+            }
+            guard ws.panels[surfaceId] != nil else {
+                result = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
+                return
+            }
+            TerminalNotificationStore.shared.addNotification(
+                tabId: ws.id,
+                surfaceId: surfaceId,
+                title: title,
+                subtitle: subtitle,
+                body: body
+            )
+            result = .ok(["workspace_id": ws.id.uuidString, "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id), "surface_id": surfaceId.uuidString, "surface_ref": v2Ref(kind: .surface, uuid: surfaceId), "window_id": v2OrNull(v2ResolveWindowId(tabManager: tabManager)?.uuidString), "window_ref": v2Ref(kind: .window, uuid: v2ResolveWindowId(tabManager: tabManager))])
+        }
+        return result
+    }
+
+    private func v2NotificationCreateForTarget(params: [String: Any]) -> V2CallResult {
+        guard let tabManager = v2ResolveTabManager(params: params) else {
+            return .err(code: "unavailable", message: "TabManager not available", data: nil)
+        }
+        guard let wsId = v2UUID(params, "workspace_id") else {
+            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
+        }
+        guard let surfaceId = v2UUID(params, "surface_id") else {
+            return .err(code: "invalid_params", message: "Missing or invalid surface_id", data: nil)
+        }
+
+        let title = (params["title"] as? String) ?? "Notification"
+        let subtitle = (params["subtitle"] as? String) ?? ""
+        let body = (params["body"] as? String) ?? ""
+
+        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
+        v2MainSync {
+            guard let ws = tabManager.tabs.first(where: { $0.id == wsId }) else {
+                result = .err(code: "not_found", message: "Workspace not found", data: ["workspace_id": wsId.uuidString])
+                return
+            }
+            guard ws.panels[surfaceId] != nil else {
+                result = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
+                return
+            }
+            TerminalNotificationStore.shared.addNotification(
+                tabId: ws.id,
+                surfaceId: surfaceId,
+                title: title,
+                subtitle: subtitle,
+                body: body
+            )
+            result = .ok(["workspace_id": ws.id.uuidString, "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id), "surface_id": surfaceId.uuidString, "surface_ref": v2Ref(kind: .surface, uuid: surfaceId), "window_id": v2OrNull(v2ResolveWindowId(tabManager: tabManager)?.uuidString), "window_ref": v2Ref(kind: .window, uuid: v2ResolveWindowId(tabManager: tabManager))])
+        }
+        return result
+    }
+
+    private func v2NotificationList() -> [String: Any] {
+        var items: [[String: Any]] = []
+        v2MainSync {
+            items = TerminalNotificationStore.shared.notifications.map { n in
+                return [
+                    "id": n.id.uuidString,
+                    "workspace_id": n.tabId.uuidString,
+                    "surface_id": v2OrNull(n.surfaceId?.uuidString),
+                    "is_read": n.isRead,
+                    "title": n.title,
+                    "subtitle": n.subtitle,
+                    "body": n.body
+                ]
+            }
+        }
+        return ["notifications": items]
+    }
+
+    private func v2NotificationClear() -> V2CallResult {
+        v2MainSync {
+            TerminalNotificationStore.shared.clearAll()
+        }
+        return .ok([:])
+    }
+}
