+++ Sources/TerminalController.swift
@@ -2808,34 +2808,7 @@ class TerminalController {
 
         return withSocketCommandPolicy(commandKey: method, isV2: true) {
             switch method {
-        case "system.ping":
-            // is_terminating_app: C11-24 — exposed so the claude-hook
-            // CLI can query "is the app shutting down?" in 250 ms before
-            // deciding whether to skip-clear during SessionEnd. Read off
-            // the AppDelegate (single shared instance, BoolGet semantics
-            // — no synchronisation needed since the field flips once
-            // and never back).
-            return v2Ok(id: id, result: [
-                "pong": true,
-                "is_terminating_app": (AppDelegate.shared?.isTerminatingApp ?? false)
-            ])
-        case "system.capabilities":
-            return v2Ok(id: id, result: v2Capabilities())
 
-        case "system.identify":
-            return v2Ok(id: id, result: v2Identify(params: params))
-        case "system.brand":
-            return v2Ok(id: id, result: v2SystemBrand())
-        case "system.tree":
-            return v2Result(id: id, self.v2SystemTree(params: params))
-        case "auth.login":
-            return v2Ok(
-                id: id,
-                result: [
-                    "authenticated": true,
-                    "required": accessMode.requiresPasswordAuth
-                ]
-            )
 
         // Windows — see Sources/SocketHandlers/WindowHandlers.swift (C11-159)
 
@@ -2894,48 +2867,12 @@ class TerminalController {
             return v2Result(id: id, self.v2WorkspaceParseBlueprint(params: params))
 
         // Snapshots (CMUX-37 Phase 1)
-        case "snapshot.create":
-            return v2Result(id: id, self.v2SnapshotCreate(params: params))
-        case "snapshot.restore":
-            return v2Result(id: id, self.v2SnapshotRestore(params: params))
-        case "snapshot.restore_set":
-            return v2Result(id: id, self.v2SnapshotRestoreSet(params: params))
-        case "snapshot.list":
-            return v2Result(id: id, self.v2SnapshotList(params: params))
-        case "snapshot.list_sets":
-            return v2Result(id: id, self.v2SnapshotListSets(params: params))
 
         // Themes (CMUX-35)
-        case "theme.list":
-            return v2Result(id: id, .ok(ThemeSocketMethods.list()))
-        case "theme.get":
-            return v2Result(id: id, .ok(ThemeSocketMethods.get(params: params)))
-        case "theme.set_active":
-            return v2Result(id: id, .ok(ThemeSocketMethods.setActive(params: params)))
-        case "theme.clear_active":
-            return v2Result(id: id, .ok(ThemeSocketMethods.clearActive()))
-        case "theme.reload":
-            return v2Result(id: id, .ok(ThemeSocketMethods.reload()))
-        case "theme.paths":
-            return v2Result(id: id, .ok(ThemeSocketMethods.paths()))
-        case "theme.dump":
-            return v2Result(id: id, .ok(ThemeSocketMethods.dumpActive(params: params)))
-        case "theme.validate":
-            return v2Result(id: id, .ok(ThemeSocketMethods.validate(params: params)))
-        case "theme.diff":
-            return v2Result(id: id, .ok(ThemeSocketMethods.diff(params: params)))
-        case "theme.inherit":
-            return v2Result(id: id, .ok(ThemeSocketMethods.inherit(params: params)))
 
         // Settings
-        case "settings.open":
-            return v2Result(id: id, self.v2SettingsOpen(params: params))
 
         // Feedback
-        case "feedback.open":
-            return v2Result(id: id, self.v2FeedbackOpen(params: params))
-        case "feedback.submit":
-            return v2Result(id: id, self.v2FeedbackSubmit(params: params))
 
 
         // Surfaces / input
@@ -2959,8 +2896,6 @@ class TerminalController {
             return v2Result(id: id, self.v2SurfaceReorder(params: params))
         case "surface.action":
             return v2Result(id: id, self.v2TabAction(params: params))
-        case "tab.action":
-            return v2Result(id: id, self.v2TabAction(params: params))
         case "surface.drag_to_split":
             return v2Result(id: id, self.v2SurfaceDragToSplit(params: params))
         case "surface.refresh":
@@ -2970,10 +2905,6 @@ class TerminalController {
         case "debug.terminals":
             return v2Result(id: id, self.v2DebugTerminals(params: params))
         // C11-131: explicit operator-grade state verbs.
-        case "session.save":
-            return v2Result(id: id, self.v2SessionSave(params: params))
-        case "app.restart":
-            return v2Result(id: id, self.v2AppRestart(params: params))
 #if DEBUG
         case "debug.session.save_and_load":
             return v2Result(id: id, self.v2DebugSessionSaveAndLoad(params: params))
@@ -2989,18 +2920,6 @@ class TerminalController {
         case "surface.trigger_flash":
             return v2Result(id: id, self.v2SurfaceTriggerFlash(params: params))
         // C11-24 conversation store
-        case "conversation.claim":
-            return v2Result(id: id, self.v2ConversationClaim(params: params))
-        case "conversation.push":
-            return v2Result(id: id, self.v2ConversationPush(params: params))
-        case "conversation.tombstone":
-            return v2Result(id: id, self.v2ConversationTombstone(params: params))
-        case "conversation.list":
-            return v2Result(id: id, self.v2ConversationList(params: params))
-        case "conversation.get":
-            return v2Result(id: id, self.v2ConversationGet(params: params))
-        case "conversation.clear":
-            return v2Result(id: id, self.v2ConversationClear(params: params))
         case "surface.cancel_flash":
             return v2Result(id: id, self.v2SurfaceCancelFlash(params: params))
         case "surface.set_metadata":
@@ -3011,8 +2930,6 @@ class TerminalController {
             return v2Result(id: id, self.v2SurfaceClearMetadata(params: params))
 
         // Mailbox
-        case "mailbox.resolve":
-            return v2Result(id: id, self.v2MailboxResolve(params: params))
 
         // Panes
         case "pane.list":
@@ -3043,22 +2960,8 @@ class TerminalController {
             return v2Result(id: id, self.v2PaneClearMetadata(params: params))
 
         // Notifications
-        case "notification.create":
-            return v2Result(id: id, self.v2NotificationCreate(params: params))
-        case "notification.create_for_surface":
-            return v2Result(id: id, self.v2NotificationCreateForSurface(params: params))
-        case "notification.create_for_target":
-            return v2Result(id: id, self.v2NotificationCreateForTarget(params: params))
-        case "notification.list":
-            return v2Ok(id: id, result: self.v2NotificationList())
-        case "notification.clear":
-            return v2Result(id: id, self.v2NotificationClear())
 
         // App focus
-        case "app.focus_override.set":
-            return v2Result(id: id, self.v2AppFocusOverride(params: params))
-        case "app.simulate_active":
-            return v2Result(id: id, self.v2AppSimulateActive())
 
         // Browser
         case "browser.open_split":
@@ -3231,14 +3134,8 @@ class TerminalController {
             return v2Result(id: id, self.v2BrowserInputTouch(params: params))
 
         // Markdown
-        case "markdown.open":
-            return v2Result(id: id, self.v2MarkdownOpen(params: params))
-        case "markdown.get_content":
-            return v2Result(id: id, self.v2MarkdownGetContent(params: params))
 
         // M3 — structured sidebar state (extends v1 `sidebar_state` with agent_chip)
-        case "sidebar.state":
-            return v2Result(id: id, self.v2SidebarState(params: params))
 
         // surface.read_text is on the socketWorker execution policy and is
         // dispatched by socketWorkerV2Response.
@@ -3342,515 +3239,19 @@ class TerminalController {
     /// unchanged. Runs on the main actor exactly like the switch it replaces.
     func v2DispatchExtracted(_ method: String, id: Any?, params: [String: Any]) -> String? {
         if method.hasPrefix("window.") { return v2DispatchWindow(method, id: id, params: params) }
+        if method.hasPrefix("theme.") { return v2DispatchTheme(method, id: id, params: params) }
+        if method.hasPrefix("system.") || method.hasPrefix("auth.") || method.hasPrefix("app.") { return v2DispatchSystem(method, id: id, params: params) }
+        if method.hasPrefix("snapshot.") { return v2DispatchSnapshot(method, id: id, params: params) }
+        if method.hasPrefix("conversation.") { return v2DispatchConversation(method, id: id, params: params) }
+        if method.hasPrefix("notification.") { return v2DispatchNotification(method, id: id, params: params) }
+        if method.hasPrefix("markdown.") || method.hasPrefix("feedback.") { return v2DispatchMarkdownFeedback(method, id: id, params: params) }
+        if method.hasPrefix("settings.") || method.hasPrefix("sidebar.") || method.hasPrefix("session.") || method.hasPrefix("tab.") || method.hasPrefix("mailbox.") { return v2DispatchMisc(method, id: id, params: params) }
         return nil
     }
 
-    private func v2Capabilities() -> [String: Any] {
-        var methods: [String] = [
-            "system.ping",
-            "system.capabilities",
-            "system.identify",
-            "system.brand",
-            "system.tree",
-            "auth.login",
-            "window.list",
-            "window.current",
-            "window.focus",
-            "window.create",
-            "window.close",
-            "workspace.list",
-            "workspace.create",
-            "workspace.select",
-            "workspace.current",
-            "workspace.close",
-            "workspace.move_to_window",
-            "workspace.reorder",
-            "workspace.rename",
-            "workspace.action",
-            "workspace.next",
-            "workspace.previous",
-            "workspace.last",
-            "workspace.remote.configure",
-            "workspace.remote.reconnect",
-            "workspace.remote.disconnect",
-            "workspace.remote.status",
-            "workspace.remote.terminal_session_end",
-            "workspace.set_metadata",
-            "workspace.get_metadata",
-            "workspace.clear_metadata",
-            "workspace.apply",
-            "settings.open",
-            "feedback.open",
-            "feedback.submit",
-            "surface.list",
-            "surface.current",
-            "surface.focus",
-            "surface.split",
-            "surface.create",
-            "surface.close",
-            "surface.drag_to_split",
-            "surface.move",
-            "surface.reorder",
-            "surface.action",
-            "tab.action",
-            "surface.refresh",
-            "surface.health",
-            "debug.terminals",
-            "surface.send_text",
-            "surface.send_key",
-            "surface.read_text",
-            "surface.clear_history",
-            "surface.trigger_flash",
-            "surface.cancel_flash",
-            "surface.set_metadata",
-            "surface.get_metadata",
-            "surface.clear_metadata",
-            "mailbox.resolve",
-            "pane.list",
-            "pane.focus",
-            "pane.surfaces",
-            "pane.create",
-            "pane.resize",
-            "pane.swap",
-            "pane.break",
-            "pane.join",
-            "pane.last",
-            "pane.set_metadata",
-            "pane.get_metadata",
-            "pane.clear_metadata",
-            "notification.create",
-            "notification.create_for_surface",
-            "notification.create_for_target",
-            "notification.list",
-            "notification.clear",
-            "app.focus_override.set",
-            "app.simulate_active",
-            "markdown.open",
-            "markdown.get_content",
-            "sidebar.state",
-            "browser.open_split",
-            "browser.navigate",
-            "browser.back",
-            "browser.forward",
-            "browser.reload",
-            "browser.url.get",
-            "browser.snapshot",
-            "browser.eval",
-            "browser.wait",
-            "browser.click",
-            "browser.dblclick",
-            "browser.hover",
-            "browser.focus",
-            "browser.type",
-            "browser.fill",
-            "browser.press",
-            "browser.keydown",
-            "browser.keyup",
-            "browser.check",
-            "browser.uncheck",
-            "browser.select",
-            "browser.scroll",
-            "browser.scroll_into_view",
-            "browser.screenshot",
-            "browser.get.text",
-            "browser.get.html",
-            "browser.get.value",
-            "browser.get.attr",
-            "browser.get.title",
-            "browser.get.count",
-            "browser.get.box",
-            "browser.get.styles",
-            "browser.is.visible",
-            "browser.is.enabled",
-            "browser.is.checked",
-            "browser.focus_webview",
-            "browser.is_webview_focused",
-            "browser.find.role",
-            "browser.find.text",
-            "browser.find.label",
-            "browser.find.placeholder",
-            "browser.find.alt",
-            "browser.find.title",
-            "browser.find.testid",
-            "browser.find.first",
-            "browser.find.last",
-            "browser.find.nth",
-            "browser.frame.select",
-            "browser.frame.main",
-            "browser.dialog.accept",
-            "browser.dialog.dismiss",
-            "browser.download.wait",
-            "browser.cookies.get",
-            "browser.cookies.set",
-            "browser.cookies.clear",
-            "browser.storage.get",
-            "browser.storage.set",
-            "browser.storage.clear",
-            "browser.tab.new",
-            "browser.tab.list",
-            "browser.tab.switch",
-            "browser.tab.close",
-            "browser.console.list",
-            "browser.console.clear",
-            "browser.errors.list",
-            "browser.highlight",
-            "browser.state.save",
-            "browser.state.load",
-            "browser.addinitscript",
-            "browser.addscript",
-            "browser.addstyle",
-            "browser.viewport.set",
-            "browser.geolocation.set",
-            "browser.offline.set",
-            "browser.trace.start",
-            "browser.trace.stop",
-            "browser.network.route",
-            "browser.network.unroute",
-            "browser.network.requests",
-            "browser.screencast.start",
-            "browser.screencast.stop",
-            "browser.input_mouse",
-            "browser.input_keyboard",
-            "browser.input_touch",
-        ]
-#if DEBUG
-        methods.append(contentsOf: [
-            "debug.shortcut.set",
-            "debug.shortcut.simulate",
-            "debug.type",
-            "debug.app.activate",
-            "debug.command_palette.toggle",
-            "debug.command_palette.rename_tab.open",
-            "debug.command_palette.visible",
-            "debug.command_palette.selection",
-            "debug.command_palette.results",
-            "debug.command_palette.rename_input.interact",
-            "debug.command_palette.rename_input.delete_backward",
-            "debug.command_palette.rename_input.selection",
-            "debug.command_palette.rename_input.select_all",
-            "debug.browser.address_bar_focused",
-            "debug.browser.favicon",
-            "debug.sidebar.visible",
-            "debug.terminal.is_focused",
-            "debug.terminal.read_text",
-            "debug.terminal.render_stats",
-            "debug.layout",
-            "debug.portal.stats",
-            "debug.bonsplit_underflow.count",
-            "debug.bonsplit_underflow.reset",
-            "debug.empty_panel.count",
-            "debug.empty_panel.reset",
-            "debug.notification.focus",
-            "debug.flash.count",
-            "debug.flash.reset",
-            "debug.panel_snapshot",
-            "debug.panel_snapshot.reset",
-            "debug.window.screenshot",
-            "debug.session.round_trip",
-            "debug.session.round_trip_workspaces",
-        ])
-#endif
 
-        return [
-            "protocol": "cmux-socket",
-            "version": 2,
-            "socket_path": socketPath,
-            "access_mode": accessMode.rawValue,
-            "methods": methods.sorted()
-        ]
-    }
 
-    private func v2Identify(params: [String: Any]) -> [String: Any] {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return [
-                "socket_path": socketPath,
-                "focused": NSNull(),
-                "caller": NSNull()
-            ]
-        }
 
-        var focused: [String: Any] = [:]
-        v2MainSync {
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            if let wsId = tabManager.selectedTabId,
-               let ws = tabManager.tabs.first(where: { $0.id == wsId }) {
-                let paneUUID = ws.bonsplitController.focusedPaneId?.id
-                let surfaceUUID = ws.focusedPanelId
-                focused = [
-                    "window_id": v2OrNull(windowId?.uuidString),
-                    "window_ref": v2Ref(kind: .window, uuid: windowId),
-                    "workspace_id": wsId.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: wsId),
-                    "pane_id": v2OrNull(paneUUID?.uuidString),
-                    "pane_ref": v2Ref(kind: .pane, uuid: paneUUID),
-                    "surface_id": v2OrNull(surfaceUUID?.uuidString),
-                    "surface_ref": v2Ref(kind: .surface, uuid: surfaceUUID),
-                    "tab_id": v2OrNull(surfaceUUID?.uuidString),
-                    "tab_ref": v2TabRef(uuid: surfaceUUID),
-                    "surface_type": v2OrNull(surfaceUUID.flatMap { ws.panels[$0]?.panelType.rawValue }),
-                    "is_browser_surface": v2OrNull(surfaceUUID.flatMap { ws.panels[$0]?.panelType == .browser })
-                ]
-            } else {
-                focused = [
-                    "window_id": v2OrNull(windowId?.uuidString),
-                    "window_ref": v2Ref(kind: .window, uuid: windowId)
-                ]
-            }
-        }
-
-        // Optionally validate a caller-provided location (useful for agents calling from inside a surface).
-        var resolvedCaller: [String: Any]? = nil
-        if let callerObj = params["caller"] as? [String: Any],
-           let wsId = v2UUIDAny(callerObj["workspace_id"]) {
-            let surfaceId = v2UUIDAny(callerObj["surface_id"]) ?? v2UUIDAny(callerObj["tab_id"])
-            v2MainSync {
-                let callerTabManager = AppDelegate.shared?.tabManagerFor(tabId: wsId) ?? tabManager
-                if let ws = callerTabManager.tabs.first(where: { $0.id == wsId }) {
-                    let callerWindowId = v2ResolveWindowId(tabManager: callerTabManager)
-                    var payload: [String: Any] = [
-                        "window_id": v2OrNull(callerWindowId?.uuidString),
-                        "window_ref": v2Ref(kind: .window, uuid: callerWindowId),
-                        "workspace_id": wsId.uuidString,
-                        "workspace_ref": v2Ref(kind: .workspace, uuid: wsId)
-                    ]
-
-                    if let surfaceId, ws.panels[surfaceId] != nil {
-                        let paneUUID = ws.paneId(forPanelId: surfaceId)?.id
-                        payload["surface_id"] = surfaceId.uuidString
-                        payload["surface_ref"] = v2Ref(kind: .surface, uuid: surfaceId)
-                        payload["tab_id"] = surfaceId.uuidString
-                        payload["tab_ref"] = v2TabRef(uuid: surfaceId)
-                        payload["surface_type"] = v2OrNull(ws.panels[surfaceId]?.panelType.rawValue)
-                        payload["is_browser_surface"] = v2OrNull(ws.panels[surfaceId]?.panelType == .browser)
-                        payload["pane_id"] = v2OrNull(paneUUID?.uuidString)
-                        payload["pane_ref"] = v2Ref(kind: .pane, uuid: paneUUID)
-                    } else {
-                        payload["surface_id"] = NSNull()
-                        payload["surface_ref"] = NSNull()
-                        payload["tab_id"] = NSNull()
-                        payload["tab_ref"] = NSNull()
-                        payload["surface_type"] = NSNull()
-                        payload["is_browser_surface"] = NSNull()
-                        payload["pane_id"] = NSNull()
-                        payload["pane_ref"] = NSNull()
-                    }
-                    resolvedCaller = payload
-                }
-            }
-        }
-
-        return [
-            "socket_path": socketPath,
-            "focused": focused.isEmpty ? NSNull() : focused,
-            "caller": v2OrNull(resolvedCaller)
-        ]
-    }
-
-    private func v2SystemBrand() -> [String: Any] {
-        // Bundle info is read from the running process — verifies the built
-        // plist, not the source tree. Matches socket threading policy: this
-        // is a fast snapshot with no UI mutation.
-        let infoDict: [String: Any] = Bundle.main.infoDictionary ?? [:]
-        let bundleIdentifier = (Bundle.main.bundleIdentifier ?? "").trimmingCharacters(in: .whitespacesAndNewlines)
-        let displayName = (infoDict["CFBundleDisplayName"] as? String) ?? (infoDict["CFBundleName"] as? String) ?? ""
-        let bundleName = (infoDict["CFBundleName"] as? String) ?? displayName
-        let iconName = (infoDict["CFBundleIconName"] as? String)
-            ?? (infoDict["CFBundleIconFile"] as? String)
-            ?? "AppIcon"
-        let shortVersion = (infoDict["CFBundleShortVersionString"] as? String) ?? ""
-        let build = (infoDict["CFBundleVersion"] as? String) ?? ""
-
-        let channel: String
-        if bundleIdentifier.contains(".debug") {
-            channel = "dev"
-        } else if bundleIdentifier.hasSuffix(".nightly") {
-            channel = "nightly"
-        } else if bundleIdentifier.hasSuffix(".staging") {
-            channel = "staging"
-        } else {
-            channel = "stable"
-        }
-
-        var palette: [String: String] = [:]
-        for (key, hex) in BrandColors.paletteHex {
-            palette[key] = hex
-        }
-
-        return [
-            "channel": channel,
-            "bundle": [
-                "identifier": bundleIdentifier,
-                "display_name": displayName,
-                "name": bundleName,
-                "icon_name": iconName,
-                "short_version": shortVersion,
-                "build": build
-            ],
-            "palette": palette,
-            "accent_hex": BrandColors.goldHex,
-            "font_family": BrandColors.fontFamily
-        ]
-    }
-
-    private func v2SystemTree(params: [String: Any]) -> V2CallResult {
-        let workspaceFilter = v2UUID(params, "workspace_id")
-        if params["workspace_id"] != nil && workspaceFilter == nil {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-
-        // Resolve scope: explicit `scope` wins, then legacy `all_windows`, then default workspace.
-        // Legal scope values: "workspace" | "window" | "all".
-        let rawScope = (params["scope"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        let scope: String
-        if let rawScope, !rawScope.isEmpty {
-            switch rawScope {
-            case "workspace", "window", "all":
-                scope = rawScope
-            default:
-                return .err(
-                    code: "invalid_params",
-                    message: "Invalid scope: \(rawScope) (expected workspace|window|all)",
-                    data: ["scope": rawScope]
-                )
-            }
-        } else if let legacy = v2Bool(params, "all_windows") {
-            scope = legacy ? "all" : "window"
-        } else if workspaceFilter != nil {
-            scope = "workspace"
-        } else {
-            scope = "workspace"
-        }
-
-        var identifyParams: [String: Any] = [:]
-        if let caller = params["caller"] as? [String: Any], !caller.isEmpty {
-            identifyParams["caller"] = caller
-        }
-        let identifyPayload = v2Identify(params: identifyParams)
-        let focused = identifyPayload["focused"] as? [String: Any] ?? [:]
-        let caller = identifyPayload["caller"] as? [String: Any] ?? [:]
-        let focusedWindowId = v2UUIDAny(focused["window_id"]) ?? v2UUIDAny(focused["window_ref"])
-        let focusedWorkspaceId = v2UUIDAny(focused["workspace_id"]) ?? v2UUIDAny(focused["workspace_ref"])
-        let callerWorkspaceId = v2UUIDAny(caller["workspace_id"]) ?? v2UUIDAny(caller["workspace_ref"])
-
-        var windowNodes: [[String: Any]] = []
-        var workspaceFound = (workspaceFilter == nil)
-
-        v2MainSync {
-            guard let app = AppDelegate.shared else { return }
-            let summaries = app.listMainWindowSummaries()
-            let defaultWindowId = focusedWindowId ?? summaries.first?.windowId
-
-            // Caller-current-workspace lookup for `scope == "workspace"`. Priority:
-            //   1. Explicit --workspace (handled below by workspaceFilter branch)
-            //   2. Caller env (CMUX_WORKSPACE_ID, propagated via params.caller)
-            //   3. Focused workspace
-            //   4. Selected workspace of the current window
-            let callerScopeWorkspaceId: UUID? = callerWorkspaceId ?? focusedWorkspaceId
-
-            for (windowIndex, summary) in summaries.enumerated() {
-                guard let manager = app.tabManagerFor(windowId: summary.windowId) else { continue }
-
-                if let workspaceFilter {
-                    guard let workspaceIndex = manager.tabs.firstIndex(where: { $0.id == workspaceFilter }) else {
-                        continue
-                    }
-                    let workspace = manager.tabs[workspaceIndex]
-                    let workspaceNode = v2TreeWorkspaceNode(
-                        workspace: workspace,
-                        index: workspaceIndex,
-                        selected: workspace.id == manager.selectedTabId
-                    )
-                    windowNodes = [
-                        v2TreeWindowNode(
-                            summary: summary,
-                            index: windowIndex,
-                            workspaceNodes: [workspaceNode]
-                        )
-                    ]
-                    workspaceFound = true
-                    break
-                }
-
-                switch scope {
-                case "all":
-                    let workspaceNodesForWindow = manager.tabs.enumerated().map { workspaceIndex, workspace in
-                        v2TreeWorkspaceNode(
-                            workspace: workspace,
-                            index: workspaceIndex,
-                            selected: workspace.id == manager.selectedTabId
-                        )
-                    }
-                    windowNodes.append(
-                        v2TreeWindowNode(
-                            summary: summary,
-                            index: windowIndex,
-                            workspaceNodes: workspaceNodesForWindow
-                        )
-                    )
-
-                case "window":
-                    if summary.windowId != defaultWindowId { continue }
-                    let workspaceNodesForWindow = manager.tabs.enumerated().map { workspaceIndex, workspace in
-                        v2TreeWorkspaceNode(
-                            workspace: workspace,
-                            index: workspaceIndex,
-                            selected: workspace.id == manager.selectedTabId
-                        )
-                    }
-                    windowNodes.append(
-                        v2TreeWindowNode(
-                            summary: summary,
-                            index: windowIndex,
-                            workspaceNodes: workspaceNodesForWindow
-                        )
-                    )
-
-                case "workspace":
-                    // Find the caller's current workspace; only include the window that owns it.
-                    let target: UUID? = callerScopeWorkspaceId ?? manager.selectedTabId
-                    guard let targetId = target,
-                          let workspaceIndex = manager.tabs.firstIndex(where: { $0.id == targetId }) else {
-                        continue
-                    }
-                    let workspace = manager.tabs[workspaceIndex]
-                    let workspaceNode = v2TreeWorkspaceNode(
-                        workspace: workspace,
-                        index: workspaceIndex,
-                        selected: workspace.id == manager.selectedTabId
-                    )
-                    windowNodes = [
-                        v2TreeWindowNode(
-                            summary: summary,
-                            index: windowIndex,
-                            workspaceNodes: [workspaceNode]
-                        )
-                    ]
-                    return  // Only one workspace per "workspace" scope; exit the v2MainSync closure.
-
-                default:
-                    continue
-                }
-            }
-        }
-
-        if let workspaceFilter, !workspaceFound {
-            return .err(
-                code: "not_found",
-                message: "Workspace not found",
-                data: [
-                    "workspace_id": workspaceFilter.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceFilter)
-                ]
-            )
-        }
-
-        return .ok([
-            "active": focused.isEmpty ? (NSNull() as Any) : focused,
-            "caller": caller.isEmpty ? (NSNull() as Any) : caller,
-            "scope": scope,
-            "windows": windowNodes
-        ])
-    }
 
     private func v2TreeWindowNode(
         summary: AppDelegate.MainWindowSummary,
@@ -5597,567 +4998,10 @@ class TerminalController {
 
     // MARK: - V2 Snapshot Methods (CMUX-37 Phase 1)
 
-    /// `snapshot.create`: capture the live workspace to a `WorkspaceSnapshotFile`
-    /// on disk. Params: `workspace_id` / `surface_id` (defaults to current).
-    /// Returns `{snapshot_id, path, surface_count, workspace_ref}`.
-    ///
-    /// Socket-initiated captures **always** land in
-    /// `WorkspaceSnapshotStore.defaultDirectory()` — any caller-supplied
-    /// `params["path"]` is rejected with `invalid_params`. The CLI's
-    /// `c11 snapshot --out <path>` path is fine because the CLI process
-    /// holds the caller's real permissions; the threat here is an agent
-    /// with only socket access turning `snapshot.create` into an
-    /// arbitrary-file-write primitive (overwriting
-    /// `~/.claude/settings.json` etc.).
-    private func v2SnapshotCreate(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        if let rawPath = params["path"] as? String,
-           !rawPath.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            return .err(
-                code: "invalid_params",
-                message: "socket-initiated snapshot.create does not accept 'path'; "
-                    + "use 'c11 snapshot --out <path>' from the CLI instead",
-                data: nil
-            )
-        }
-        let originRaw = (params["origin"] as? String) ?? "manual"
-        let origin: WorkspaceSnapshotFile.Origin =
-            (originRaw == "auto-restart") ? .autoRestart : .manual
 
-        // --all: capture every open workspace, write per-workspace files
-        // AND a manifest under `~/.c11-snapshots/sets/<set_id>.json` that
-        // points at the inner ids. The manifest is the discoverable handle
-        // for `c11 restore <set_id>` — it carries no plan data of its own,
-        // so each inner snapshot stays independently restorable.
-        let captureAll = (params["all"] as? Bool) == true
-        if captureAll {
-            var snapshots: [(envelope: WorkspaceSnapshotFile, ref: String, isSelected: Bool)] = []
-            v2MainSync {
-                let source = LiveWorkspaceSnapshotSource(tabManager: tabManager)
-                let selectedId = tabManager.selectedWorkspace?.id
-                for ws in tabManager.tabs {
-                    if let envelope = source.capture(
-                        workspaceId: ws.id, origin: origin, clock: { Date() }
-                    ) {
-                        let ref = self.v2EnsureHandleRef(kind: .workspace, uuid: ws.id)
-                        snapshots.append((envelope, ref, ws.id == selectedId))
-                    }
-                }
-            }
-            let store = WorkspaceSnapshotStore()
-            var results: [[String: Any]] = []
-            var entries: [WorkspaceSnapshotSetFile.Entry] = []
-            var selectedIndex: Int?
-            var anyWriteFailed = false
-            for (offset, item) in snapshots.enumerated() {
-                let (envelope, ref, isSelected) = item
-                do {
-                    let path = try store.writeToDefaultDirectory(envelope)
-                    var row: [String: Any] = [
-                        "snapshot_id": envelope.snapshotId,
-                        "path": path.path,
-                        "surface_count": envelope.plan.surfaces.count,
-                        "workspace_ref": ref
-                    ]
-                    if isSelected { row["selected"] = true }
-                    results.append(row)
-                    entries.append(WorkspaceSnapshotSetFile.Entry(
-                        workspaceRef: ref,
-                        snapshotId: envelope.snapshotId,
-                        order: offset,
-                        selected: isSelected
-                    ))
-                    if isSelected { selectedIndex = entries.count - 1 }
-                } catch {
-                    results.append([
-                        "snapshot_id": envelope.snapshotId,
-                        "error": "\(error)",
-                        "workspace_ref": ref
-                    ])
-                    anyWriteFailed = true
-                }
-            }
-            // Build the manifest only when at least one inner snapshot
-            // wrote successfully — an all-failed run has nothing to point
-            // at. When some inner writes failed but others succeeded the
-            // manifest still ships, listing only the successful entries
-            // (the failed entries are visible via the per-snapshot result
-            // rows).
-            var payload: [String: Any] = ["snapshots": results]
-            if !entries.isEmpty {
-                let setId = WorkspaceSnapshotID.generate()
-                let manifest = WorkspaceSnapshotSetFile(
-                    version: 1,
-                    setId: setId,
-                    createdAt: Date(),
-                    c11Version: LiveWorkspaceSnapshotSource.defaultVersionString(),
-                    selectedWorkspaceIndex: selectedIndex,
-                    snapshots: entries
-                )
-                do {
-                    let setPath = try store.writeSet(manifest)
-                    payload["set_id"] = setId
-                    payload["set_path"] = setPath.path
-                } catch {
-                    // Surface the manifest-write failure but keep the
-                    // per-workspace data the caller already has.
-                    payload["set_error"] = "\(error)"
-                }
-            } else if anyWriteFailed {
-                payload["set_error"] = "no inner snapshots wrote successfully"
-            }
-            return .ok(payload)
-        }
 
-        var snapshot: WorkspaceSnapshotFile?
-        var workspaceRef = ""
-        v2MainSync {
-            guard let workspace = v2ResolveWorkspace(params: params, tabManager: tabManager) else { return }
-            let source = LiveWorkspaceSnapshotSource(tabManager: tabManager)
-            snapshot = source.capture(workspaceId: workspace.id, origin: origin, clock: { Date() })
-            workspaceRef = self.v2EnsureHandleRef(kind: .workspace, uuid: workspace.id)
-        }
-        guard let envelope = snapshot else {
-            return .err(code: "not_found", message: "Workspace not found for snapshot.create", data: nil)
-        }
-        let store = WorkspaceSnapshotStore()
-        let path: URL
-        do {
-            path = try store.writeToDefaultDirectory(envelope)
-        } catch let err as WorkspaceSnapshotStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "snapshot_write_failed", message: "\(error)", data: nil)
-        }
-        let payload: [String: Any] = [
-            "snapshot_id": envelope.snapshotId,
-            "path": path.path,
-            "surface_count": envelope.plan.surfaces.count,
-            "workspace_ref": workspaceRef
-        ]
-        return .ok(payload)
-    }
 
-    /// `snapshot.restore`: read a snapshot by id, run the embedded plan
-    /// through `WorkspaceLayoutExecutor`, optionally threading a named
-    /// restart registry (`"phase1"` → `AgentRestartRegistry.phase1`) so cc
-    /// terminals resume via `cc --resume <session-id>`. Returns the same
-    /// `ApplyResult` shape `workspace.apply` returns.
-    ///
-    /// Socket-initiated restores resolve `snapshot_id` through
-    /// `WorkspaceSnapshotStore.resolvePath(byId:)`, which validates the id
-    /// grammar and asserts the resolved realpath lives under a configured
-    /// snapshot root. Caller-supplied `params["path"]` is rejected with
-    /// `invalid_params`: an agent with socket access must not be able to
-    /// coerce the restore path-reader into parsing arbitrary files
-    /// (`/etc/passwd.json`, say) — parser errors can leak file contents.
-    /// The CLI's `c11 restore <path-or-id>` classifies locally before
-    /// sending and always submits via `snapshot_id`.
-    private func v2SnapshotRestore(params: [String: Any]) -> V2CallResult {
-        if let rawPath = params["path"] as? String,
-           !rawPath.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            return .err(
-                code: "invalid_params",
-                message: "socket-initiated snapshot.restore does not accept 'path'; "
-                    + "use 'snapshot_id' (resolved against the snapshot roots) instead",
-                data: nil
-            )
-        }
-        let rawId = (params["snapshot_id"] as? String).map {
-            $0.trimmingCharacters(in: .whitespacesAndNewlines)
-        }
-        let store = WorkspaceSnapshotStore()
-        let envelope: WorkspaceSnapshotFile
-        do {
-            if let id = rawId, !id.isEmpty {
-                envelope = try store.read(byId: id)
-            } else {
-                return .err(
-                    code: "invalid_params",
-                    message: "snapshot.restore requires 'snapshot_id'",
-                    data: nil
-                )
-            }
-        } catch let err as WorkspaceSnapshotStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "snapshot_read_failed", message: "\(error)", data: nil)
-        }
 
-        // Envelope → plan, off-main.
-        let planResult = WorkspaceSnapshotConverter.applyPlan(from: envelope)
-        let plan: WorkspaceApplyPlan
-        switch planResult {
-        case .success(let p): plan = p
-        case .failure(let err):
-            return .err(code: err.code, message: err.message, data: nil)
-        }
-        if let validationFailure = WorkspaceLayoutExecutor.validate(plan: plan) {
-            let data: [String: Any] = [
-                "failure": [
-                    "code": validationFailure.code,
-                    "step": validationFailure.step,
-                    "message": validationFailure.message
-                ]
-            ]
-            return .err(
-                code: "invalid_params",
-                message: "snapshot plan validation failed: \(validationFailure.message)",
-                data: data
-            )
-        }
-
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        // Build options: select is subject to the focus policy; the restart
-        // registry is resolved by name on the wire so snapshot files stay
-        // forward-compatible with future Phase 5 rows.
-        var options = ApplyOptions()
-        if let selectValue = params["select"] as? Bool {
-            options.select = selectValue
-        }
-        if options.select && !v2FocusAllowed(requested: true) {
-            options.select = false
-        }
-        // Pre-apply warnings surfaced through ApplyResult.warnings below so
-        // operators see them in both JSON and the CLI's `warnings:` block.
-        var preApplyWarnings: [String] = []
-        if let registryName = params["restart_registry"] as? String {
-            let resolved = AgentRestartRegistry.named(registryName)
-            options.restartRegistry = resolved
-            if resolved == nil {
-                // Unknown wire name → fall back to Phase 0 (no synthesis) but
-                // make the silent degradation visible. Reject is the
-                // tempting alternative, but the registry is designed for
-                // Phase 5 forward compatibility — an operator submitting
-                // "phase5" against an older c11 binary should still land
-                // their restore, just without restart synthesis.
-                preApplyWarnings.append(
-                    "unknown restart_registry '\(registryName)': falling back to no-op (no cc --resume synthesis)"
-                )
-            }
-        }
-
-        // In-place restore (I2): resolve the target workspace UUID either
-        // from the `target_workspace_id` param or from the focus-policy
-        // default. When the target doesn't resolve the executor surfaces
-        // `invalid_params` via ApplyResult.failures.
-        let inPlace = (params["in_place"] as? Bool) == true
-        let inPlaceTarget: UUID? = {
-            guard inPlace else { return nil }
-            if let raw = params["target_workspace_id"] as? String,
-               let uuid = UUID(uuidString: raw) {
-                return uuid
-            }
-            return nil
-        }()
-
-        var result: ApplyResult?
-        v2MainSync {
-            let deps = WorkspaceLayoutExecutorDependencies(
-                tabManager: tabManager,
-                workspaceRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .workspace, uuid: uuid) ?? "workspace:\(uuid.uuidString)"
-                },
-                surfaceRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .surface, uuid: uuid) ?? "surface:\(uuid.uuidString)"
-                },
-                paneRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .pane, uuid: uuid) ?? "pane:\(uuid.uuidString)"
-                }
-            )
-            if inPlace, let target = inPlaceTarget {
-                result = WorkspaceLayoutExecutor.applyToExistingWorkspace(
-                    plan,
-                    options: options,
-                    dependencies: deps,
-                    existingWorkspaceId: target
-                )
-            } else if inPlace {
-                // Requested in_place but no target resolver succeeded.
-                // Synthesise the same invalid_params failure the executor
-                // would emit so callers see a consistent shape.
-                result = ApplyResult(
-                    workspaceRef: "",
-                    surfaceRefs: [:],
-                    paneRefs: [:],
-                    timings: [],
-                    warnings: ["in_place restore requires a resolvable target_workspace_id"],
-                    failures: [ApplyFailure(
-                        code: "invalid_params",
-                        step: "validate",
-                        message: "in_place restore requires 'target_workspace_id' as a UUID string"
-                    )]
-                )
-            } else {
-                result = WorkspaceLayoutExecutor.apply(plan, options: options, dependencies: deps)
-            }
-        }
-        guard var applyResult = result else {
-            return .err(code: "internal_error", message: "Executor returned no result", data: nil)
-        }
-        if !preApplyWarnings.isEmpty {
-            applyResult.warnings = preApplyWarnings + applyResult.warnings
-        }
-        do {
-            let encoded = try JSONEncoder().encode(applyResult)
-            let asAny = try JSONSerialization.jsonObject(with: encoded, options: [])
-            return .ok(asAny)
-        } catch {
-            return .err(code: "internal_error", message: "Failed to encode ApplyResult: \(error)", data: nil)
-        }
-    }
-
-    /// `snapshot.restore_set`: rehydrate every per-workspace snapshot
-    /// referenced by a `WorkspaceSnapshotSetFile` manifest. Required
-    /// param: `set_id` (resolved against
-    /// `~/.c11-snapshots/sets/<set_id>.json` only — never a
-    /// caller-supplied path). Optional: `restart_registry` (passes
-    /// through to each inner restore), `select` (best-effort focus on
-    /// the entry the manifest marked as selected, subject to the same
-    /// focus policy `snapshot.restore` honors).
-    ///
-    /// `in_place` is rejected here: a set restore creates several fresh
-    /// workspaces, and there is no single target workspace to replace.
-    /// The CLI catches this earlier; the v2 method validates again so
-    /// programmatic callers cannot trip the executor.
-    private func v2SnapshotRestoreSet(params: [String: Any]) -> V2CallResult {
-        if let rawPath = params["path"] as? String,
-           !rawPath.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            return .err(
-                code: "invalid_params",
-                message: "socket-initiated snapshot.restore_set does not accept 'path'; "
-                    + "use 'set_id' instead",
-                data: nil
-            )
-        }
-        if (params["in_place"] as? Bool) == true {
-            return .err(
-                code: "invalid_params",
-                message: "snapshot.restore_set does not support in_place; a set restore creates fresh workspaces",
-                data: nil
-            )
-        }
-        guard let rawId = (params["set_id"] as? String)?
-                .trimmingCharacters(in: .whitespacesAndNewlines),
-              !rawId.isEmpty else {
-            return .err(
-                code: "invalid_params",
-                message: "snapshot.restore_set requires 'set_id'",
-                data: nil
-            )
-        }
-        let store = WorkspaceSnapshotStore()
-        let manifest: WorkspaceSnapshotSetFile
-        do {
-            manifest = try store.readSet(byId: rawId)
-        } catch let err as WorkspaceSnapshotStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "snapshot_set_read_failed", message: "\(error)", data: nil)
-        }
-
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        // Resolve restart registry once for the whole set; missing names
-        // degrade to no-op with a per-set warning (matches snapshot.restore).
-        var preApplyWarnings: [String] = []
-        var optionsTemplate = ApplyOptions()
-        optionsTemplate.select = false  // each apply runs unselect; we re-establish at the end
-        if let registryName = params["restart_registry"] as? String {
-            let resolved = AgentRestartRegistry.named(registryName)
-            optionsTemplate.restartRegistry = resolved
-            if resolved == nil {
-                preApplyWarnings.append(
-                    "unknown restart_registry '\(registryName)': falling back to no-op (no cc --resume synthesis)"
-                )
-            }
-        }
-
-        let allowFocus = v2FocusAllowed(requested: (params["select"] as? Bool) ?? true)
-
-        var workspaceResults: [[String: Any]] = []
-        var selectedWorkspaceRef: String?
-
-        // Apply each inner snapshot sequentially. Order is taken from the
-        // manifest's `order` field as written; we sort defensively here so
-        // a future hand-edited manifest with shuffled keys still applies
-        // in capture-time order.
-        let sortedEntries = manifest.snapshots.sorted { $0.order < $1.order }
-        for entry in sortedEntries {
-            let envelope: WorkspaceSnapshotFile
-            do {
-                envelope = try store.read(byId: entry.snapshotId)
-            } catch let err as WorkspaceSnapshotStore.StoreError {
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": "\(err)",
-                    "code": err.code
-                ])
-                continue
-            } catch {
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": "\(error)"
-                ])
-                continue
-            }
-            let planResult = WorkspaceSnapshotConverter.applyPlan(from: envelope)
-            let plan: WorkspaceApplyPlan
-            switch planResult {
-            case .success(let p): plan = p
-            case .failure(let err):
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": err.message,
-                    "code": err.code
-                ])
-                continue
-            }
-            if let validationFailure = WorkspaceLayoutExecutor.validate(plan: plan) {
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": validationFailure.message,
-                    "code": validationFailure.code
-                ])
-                continue
-            }
-
-            var result: ApplyResult?
-            v2MainSync {
-                let deps = WorkspaceLayoutExecutorDependencies(
-                    tabManager: tabManager,
-                    workspaceRefMinter: { [weak self] uuid in
-                        self?.v2EnsureHandleRef(kind: .workspace, uuid: uuid) ?? "workspace:\(uuid.uuidString)"
-                    },
-                    surfaceRefMinter: { [weak self] uuid in
-                        self?.v2EnsureHandleRef(kind: .surface, uuid: uuid) ?? "surface:\(uuid.uuidString)"
-                    },
-                    paneRefMinter: { [weak self] uuid in
-                        self?.v2EnsureHandleRef(kind: .pane, uuid: uuid) ?? "pane:\(uuid.uuidString)"
-                    }
-                )
-                result = WorkspaceLayoutExecutor.apply(plan, options: optionsTemplate, dependencies: deps)
-            }
-            guard let applyResult = result else {
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": "executor returned no result"
-                ])
-                continue
-            }
-            do {
-                let encoded = try JSONEncoder().encode(applyResult)
-                if let asAny = try JSONSerialization.jsonObject(with: encoded, options: []) as? [String: Any] {
-                    var row = asAny
-                    row["snapshot_id"] = entry.snapshotId
-                    row["original_workspace_ref"] = entry.workspaceRef
-                    if entry.selected { row["selected"] = true }
-                    workspaceResults.append(row)
-                    if entry.selected, !applyResult.workspaceRef.isEmpty {
-                        selectedWorkspaceRef = applyResult.workspaceRef
-                    }
-                }
-            } catch {
-                workspaceResults.append([
-                    "snapshot_id": entry.snapshotId,
-                    "workspace_ref": entry.workspaceRef,
-                    "error": "encode failed: \(error)"
-                ])
-            }
-        }
-
-        // Re-establish selection on a best-effort basis, subject to the
-        // socket focus policy. Convert the ref back to a UUID via the
-        // existing handle table, then look up the live workspace.
-        if allowFocus, let targetRef = selectedWorkspaceRef,
-           let uuid = self.v2ResolveHandleRef(targetRef) {
-            v2MainSync {
-                if let ws = tabManager.tabs.first(where: { $0.id == uuid }) {
-                    tabManager.selectWorkspace(ws)
-                }
-            }
-        }
-
-        var payload: [String: Any] = [
-            "set_id": manifest.setId,
-            "workspaces": workspaceResults,
-            "warnings": preApplyWarnings
-        ]
-        if let ref = selectedWorkspaceRef {
-            payload["selected_workspace_ref"] = ref
-        }
-        return .ok(payload)
-    }
-
-    /// `snapshot.list_sets`: enumerate manifests under
-    /// `~/.c11-snapshots/sets/`. Returns rows sorted newest-first.
-    private func v2SnapshotListSets(params: [String: Any]) -> V2CallResult {
-        let store = WorkspaceSnapshotStore()
-        let entries: [WorkspaceSnapshotSetIndex]
-        do {
-            entries = try store.listSets()
-        } catch let err as WorkspaceSnapshotStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "snapshot_list_sets_failed", message: "\(error)", data: nil)
-        }
-        do {
-            let encoder = JSONEncoder()
-            encoder.dateEncodingStrategy = .custom { date, encoder in
-                var container = encoder.singleValueContainer()
-                try container.encode(workspaceSnapshotDateFormatter.string(from: date))
-            }
-            let encoded = try encoder.encode(entries)
-            let asAny = try JSONSerialization.jsonObject(with: encoded, options: [])
-            return .ok(["sets": asAny])
-        } catch {
-            return .err(code: "internal_error", message: "\(error)", data: nil)
-        }
-    }
-
-    /// `snapshot.list`: enumerate `~/.c11-snapshots/` + `~/.cmux-snapshots/`
-    /// and return entries sorted newest-first. Pure filesystem; runs
-    /// off-main.
-    private func v2SnapshotList(params: [String: Any]) -> V2CallResult {
-        let store = WorkspaceSnapshotStore()
-        let entries: [WorkspaceSnapshotIndex]
-        do {
-            entries = try store.list()
-        } catch let err as WorkspaceSnapshotStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "snapshot_list_failed", message: "\(error)", data: nil)
-        }
-        do {
-            let encoder = JSONEncoder()
-            // Match the store's write-side formatter (fractional seconds)
-            // so a timestamp written by `WorkspaceSnapshotStore.write` and
-            // re-read into a `snapshot.list` response round-trips
-            // bit-for-bit. The default `.iso8601` strategy drops subsecond
-            // precision and creates a write-vs-read mismatch.
-            encoder.dateEncodingStrategy = .custom { date, encoder in
-                var container = encoder.singleValueContainer()
-                try container.encode(workspaceSnapshotDateFormatter.string(from: date))
-            }
-            let encoded = try encoder.encode(entries)
-            let asAny = try JSONSerialization.jsonObject(with: encoded, options: [])
-            return .ok(["snapshots": asAny])
-        } catch {
-            return .err(code: "internal_error", message: "\(error)", data: nil)
-        }
-    }
 
     private func v2WorkspaceSetMetadata(params: [String: Any]) -> V2CallResult {
         // Parse + validate off-main per the socket command threading policy
@@ -6498,273 +5342,6 @@ class TerminalController {
         return result
     }
 
-    private func v2TabAction(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let action = v2ActionKey(params) else {
-            return .err(code: "invalid_params", message: "Missing action", data: nil)
-        }
-
-        let supportedActions = [
-            "rename", "clear_name",
-            "close_left", "close_right", "close_others",
-            "new_terminal_right", "new_browser_right",
-            "reload", "duplicate",
-            "pin", "unpin", "mark_read", "mark_unread"
-        ]
-
-        var result: V2CallResult = .err(code: "invalid_params", message: "Unknown tab action", data: [
-            "action": action,
-            "supported_actions": supportedActions
-        ])
-
-        v2MainSync {
-            guard let workspace = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-
-            let surfaceId = v2UUID(params, "surface_id") ?? v2UUID(params, "tab_id") ?? workspace.focusedPanelId
-            guard let surfaceId else {
-                result = .err(code: "not_found", message: "No focused tab", data: nil)
-                return
-            }
-            guard workspace.panels[surfaceId] != nil else {
-                result = .err(code: "not_found", message: "Tab not found", data: [
-                    "surface_id": surfaceId.uuidString,
-                    "surface_ref": v2Ref(kind: .surface, uuid: surfaceId),
-                    "tab_id": surfaceId.uuidString,
-                    "tab_ref": v2TabRef(uuid: surfaceId)
-                ])
-                return
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-
-            @MainActor
-            func finish(_ extras: [String: Any] = [:]) {
-                var payload: [String: Any] = [
-                    "action": action,
-                    "window_id": v2OrNull(windowId?.uuidString),
-                    "window_ref": v2Ref(kind: .window, uuid: windowId),
-                    "workspace_id": workspace.id.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                    "surface_id": surfaceId.uuidString,
-                    "surface_ref": v2Ref(kind: .surface, uuid: surfaceId),
-                    "tab_id": surfaceId.uuidString,
-                    "tab_ref": v2TabRef(uuid: surfaceId)
-                ]
-                if let paneId = workspace.paneId(forPanelId: surfaceId)?.id {
-                    payload["pane_id"] = paneId.uuidString
-                    payload["pane_ref"] = v2Ref(kind: .pane, uuid: paneId)
-                } else {
-                    payload["pane_id"] = NSNull()
-                    payload["pane_ref"] = NSNull()
-                }
-                for (key, value) in extras {
-                    payload[key] = value
-                }
-                result = .ok(payload)
-            }
-
-            @MainActor
-            func insertionIndexToRight(anchorTabId: TabID, inPane paneId: PaneID) -> Int {
-                let tabs = workspace.bonsplitController.tabs(inPane: paneId)
-                guard let anchorIndex = tabs.firstIndex(where: { $0.id == anchorTabId }) else { return tabs.count }
-                let pinnedCount = tabs.reduce(into: 0) { count, tab in
-                    if let panelId = workspace.panelIdFromSurfaceId(tab.id),
-                       workspace.isPanelPinned(panelId) {
-                        count += 1
-                    }
-                }
-                let rawTarget = min(anchorIndex + 1, tabs.count)
-                return max(rawTarget, pinnedCount)
-            }
-
-            @MainActor
-            func closeTabs(_ tabIds: [TabID]) -> (closed: Int, skippedPinned: Int) {
-                var closed = 0
-                var skippedPinned = 0
-                for tabId in tabIds {
-                    guard let panelId = workspace.panelIdFromSurfaceId(tabId) else { continue }
-                    if workspace.isPanelPinned(panelId) {
-                        skippedPinned += 1
-                        continue
-                    }
-                    if workspace.panels.count <= 1 {
-                        break
-                    }
-                    if workspace.closePanel(panelId, force: true) {
-                        closed += 1
-                    }
-                }
-                return (closed, skippedPinned)
-            }
-
-            switch action {
-            case "rename":
-                guard let titleRaw = v2String(params, "title"),
-                      !titleRaw.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
-                    result = .err(code: "invalid_params", message: "Missing or invalid title", data: nil)
-                    return
-                }
-                let title = titleRaw.trimmingCharacters(in: .whitespacesAndNewlines)
-                workspace.setPanelCustomTitle(panelId: surfaceId, title: title)
-                finish(["title": title])
-
-            case "clear_name":
-                workspace.setPanelCustomTitle(panelId: surfaceId, title: nil)
-                finish()
-
-            case "pin":
-                workspace.setPanelPinned(panelId: surfaceId, pinned: true)
-                finish(["pinned": true])
-
-            case "unpin":
-                workspace.setPanelPinned(panelId: surfaceId, pinned: false)
-                finish(["pinned": false])
-
-            case "mark_read":
-                workspace.markPanelRead(surfaceId)
-                finish()
-
-            case "mark_unread", "mark_as_unread":
-                workspace.markPanelUnread(surfaceId)
-                finish()
-
-            case "reload", "reload_tab":
-                guard let browserPanel = workspace.browserPanel(for: surfaceId) else {
-                    result = .err(code: "invalid_state", message: "Reload is only available for browser tabs", data: nil)
-                    return
-                }
-                browserPanel.reload()
-                finish()
-
-            case "duplicate", "duplicate_tab":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId),
-                      let browserPanel = workspace.browserPanel(for: surfaceId) else {
-                    result = .err(code: "invalid_state", message: "Duplicate is only available for browser tabs", data: nil)
-                    return
-                }
-
-                let targetIndex = insertionIndexToRight(anchorTabId: anchorTabId, inPane: paneId)
-                guard let newPanel = workspace.newBrowserSurface(
-                    inPane: paneId,
-                    url: browserPanel.currentURL,
-                    focus: true
-                ) else {
-                    result = .err(code: "internal_error", message: "Failed to duplicate tab", data: nil)
-                    return
-                }
-                _ = workspace.reorderSurface(panelId: newPanel.id, toIndex: targetIndex)
-                finish([
-                    "created_surface_id": newPanel.id.uuidString,
-                    "created_surface_ref": v2Ref(kind: .surface, uuid: newPanel.id),
-                    "created_tab_id": newPanel.id.uuidString,
-                    "created_tab_ref": v2TabRef(uuid: newPanel.id)
-                ])
-
-            case "new_terminal_right", "new_terminal_to_right", "new_terminal_tab_to_right":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId) else {
-                    result = .err(code: "not_found", message: "Tab pane not found", data: nil)
-                    return
-                }
-
-                let targetIndex = insertionIndexToRight(anchorTabId: anchorTabId, inPane: paneId)
-                guard let newPanel = workspace.newTerminalSurface(inPane: paneId, focus: true) else {
-                    result = .err(code: "internal_error", message: "Failed to create tab", data: nil)
-                    return
-                }
-                _ = workspace.reorderSurface(panelId: newPanel.id, toIndex: targetIndex)
-                finish([
-                    "created_surface_id": newPanel.id.uuidString,
-                    "created_surface_ref": v2Ref(kind: .surface, uuid: newPanel.id),
-                    "created_tab_id": newPanel.id.uuidString,
-                    "created_tab_ref": v2TabRef(uuid: newPanel.id)
-                ])
-
-            case "new_browser_right", "new_browser_to_right", "new_browser_tab_to_right":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId) else {
-                    result = .err(code: "not_found", message: "Tab pane not found", data: nil)
-                    return
-                }
-
-                let urlRaw = v2String(params, "url")
-                let url = urlRaw.flatMap { URL(string: $0) }
-                if urlRaw != nil && url == nil {
-                    result = .err(code: "invalid_params", message: "Invalid URL", data: ["url": v2OrNull(urlRaw)])
-                    return
-                }
-
-                let targetIndex = insertionIndexToRight(anchorTabId: anchorTabId, inPane: paneId)
-                guard let newPanel = workspace.newBrowserSurface(inPane: paneId, url: url, focus: true) else {
-                    result = .err(code: "internal_error", message: "Failed to create tab", data: nil)
-                    return
-                }
-                _ = workspace.reorderSurface(panelId: newPanel.id, toIndex: targetIndex)
-                finish([
-                    "created_surface_id": newPanel.id.uuidString,
-                    "created_surface_ref": v2Ref(kind: .surface, uuid: newPanel.id),
-                    "created_tab_id": newPanel.id.uuidString,
-                    "created_tab_ref": v2TabRef(uuid: newPanel.id)
-                ])
-
-            case "close_left", "close_to_left":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId) else {
-                    result = .err(code: "not_found", message: "Tab pane not found", data: nil)
-                    return
-                }
-                let tabs = workspace.bonsplitController.tabs(inPane: paneId)
-                guard let index = tabs.firstIndex(where: { $0.id == anchorTabId }) else {
-                    result = .err(code: "not_found", message: "Tab not found in pane", data: nil)
-                    return
-                }
-                let targetIds = Array(tabs.prefix(index).map(\.id))
-                let closeResult = closeTabs(targetIds)
-                finish(["closed": closeResult.closed, "skipped_pinned": closeResult.skippedPinned])
-
-            case "close_right", "close_to_right":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId) else {
-                    result = .err(code: "not_found", message: "Tab pane not found", data: nil)
-                    return
-                }
-                let tabs = workspace.bonsplitController.tabs(inPane: paneId)
-                guard let index = tabs.firstIndex(where: { $0.id == anchorTabId }) else {
-                    result = .err(code: "not_found", message: "Tab not found in pane", data: nil)
-                    return
-                }
-                let targetIds = (index + 1 < tabs.count) ? Array(tabs.suffix(from: index + 1).map(\.id)) : []
-                let closeResult = closeTabs(targetIds)
-                finish(["closed": closeResult.closed, "skipped_pinned": closeResult.skippedPinned])
-
-            case "close_others", "close_other_tabs":
-                guard let anchorTabId = workspace.surfaceIdFromPanelId(surfaceId),
-                      let paneId = workspace.paneId(forPanelId: surfaceId) else {
-                    result = .err(code: "not_found", message: "Tab pane not found", data: nil)
-                    return
-                }
-                let targetIds = workspace.bonsplitController.tabs(inPane: paneId)
-                    .map(\.id)
-                    .filter { $0 != anchorTabId }
-                let closeResult = closeTabs(targetIds)
-                finish(["closed": closeResult.closed, "skipped_pinned": closeResult.skippedPinned])
-
-            default:
-                result = .err(code: "invalid_params", message: "Unknown tab action", data: [
-                    "action": action,
-                    "supported_actions": supportedActions
-                ])
-            }
-        }
-
-        return result
-    }
 
     // MARK: - V2 Surface Methods
 
@@ -7625,62 +6202,7 @@ class TerminalController {
         return .ok(payload)
     }
 
-    /// C11-131 `session.save` — production cousin of the DEBUG-only
-    /// `debug.session.save_and_load`. Forces a synchronous full-app session
-    /// snapshot while the app keeps running and returns the path + counts.
-    ///
-    /// Main-actor handling (`v2MainSync`) is intentional and permitted by the
-    /// socket threading policy: this is an operator-grade, low-frequency verb
-    /// that must build the snapshot from live AppKit/window state and read it
-    /// back — it is not a telemetry hot path.
-    private func v2SessionSave(params: [String: Any]) -> V2CallResult {
-        let includeScrollback = (params["include_scrollback"] as? Bool) ?? false
-        let outPath = (params["out"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines)
-        var result: AppDelegate.SessionSaveResult?
-        v2MainSync {
-            guard let app = AppDelegate.shared else { return }
-            result = app.forceSessionSave(
-                includeScrollback: includeScrollback,
-                outPath: (outPath?.isEmpty == false) ? outPath : nil
-            )
-        }
-        guard let result else {
-            return .err(code: "session_save_failed",
-                        message: "forceSessionSave returned nil (no windows or write failed)",
-                        data: nil)
-        }
-        var payload: [String: Any] = [
-            "snapshot_path": result.snapshotPath,
-            "windows": result.windows,
-            "workspaces": result.workspaces,
-            "terminal_panels": result.terminalPanels,
-            "refs": result.refs
-        ]
-        if let out = result.outPath { payload["out_path"] = out }
-        return .ok(payload)
-    }
 
-    /// C11-131 `app.restart` — clean-shutdown choreography then relaunch.
-    /// `no_resume` restores layout without typing resume commands into panes.
-    ///
-    /// Main-actor handling is intentional: it mutates app lifecycle state
-    /// (suspend refs, snapshot, sentinel, terminate). The reply flushes
-    /// before the deferred `NSApp.terminate` fires. This verb does not steal
-    /// focus beyond the operator's explicit restart intent.
-    private func v2AppRestart(params: [String: Any]) -> V2CallResult {
-        let noResume = (params["no_resume"] as? Bool) ?? false
-        var dispatched = false
-        v2MainSync {
-            guard let app = AppDelegate.shared else { return }
-            app.performCleanRestart(resume: !noResume)
-            dispatched = true
-        }
-        guard dispatched else {
-            return .err(code: "app_restart_failed",
-                        message: "AppDelegate unavailable", data: nil)
-        }
-        return .ok(["ok": true, "resume": !noResume])
-    }
 
 #if DEBUG
     /// DEBUG-only: force an on-disk session snapshot round-trip through
@@ -9016,101 +7538,6 @@ class TerminalController {
 
     // MARK: - Mailbox cross-workspace resolution
 
-    /// Resolves a mailbox recipient name to the workspace its envelope should
-    /// be delivered into, scanning every live surface across all windows.
-    /// Backs `c11 mailbox send`, which routes the envelope into the resolved
-    /// workspace's outbox so that workspace's own dispatcher delivers it.
-    ///
-    /// Params:
-    ///   * `to` (string, required) — recipient surface name.
-    ///   * `sender_workspace_id` (uuid, required) — the sender's workspace, so
-    ///     a local match takes precedence over same-named surfaces elsewhere.
-    ///   * `workspace` (string, optional) — disambiguates a name that lives in
-    ///     more than one workspace; a workspace UUID or a `workspace:*` ref.
-    ///
-    /// Returns `{ resolution: "unique"|"ambiguous"|"unresolved",
-    /// target_workspace_id?, target_workspace_ref?, surface_ids?, candidates }`.
-    private func v2MailboxResolve(params: [String: Any]) -> V2CallResult {
-        guard let to = v2String(params, "to"), !to.isEmpty else {
-            return .err(code: "invalid_to", message: "to is required", data: nil)
-        }
-        guard let senderWorkspaceId = v2UUID(params, "sender_workspace_id") else {
-            return .err(code: "invalid_sender_workspace_id", message: "sender_workspace_id must be a UUID", data: nil)
-        }
-        let qualifierStr = v2String(params, "workspace")
-
-        return v2MainSync {
-            let surfaces = AppDelegate.shared?.mailboxAddressableSurfaces() ?? []
-
-            // Translate the optional workspace qualifier (UUID or ref string)
-            // into a concrete workspace UUID. A raw UUID is honored even when
-            // that workspace currently has no addressable surface (it then
-            // resolves to `unresolved`, the honest answer). Otherwise match the
-            // qualifier against the ref strings of the live workspaces.
-            var qualifierUUID: UUID?
-            if let qualifierStr, !qualifierStr.isEmpty {
-                if let direct = UUID(uuidString: qualifierStr) {
-                    qualifierUUID = direct
-                } else {
-                    let workspaceIds = Set(surfaces.map(\.workspaceId))
-                    qualifierUUID = workspaceIds.first { wsId in
-                        if let ref = v2Ref(kind: .workspace, uuid: wsId) as? String {
-                            return ref == qualifierStr
-                        }
-                        return false
-                    }
-                    if qualifierUUID == nil {
-                        // Unknown workspace ref → nothing can match it.
-                        return .ok([
-                            "resolution": "unresolved",
-                            "candidates": self.mailboxCandidatePayload(
-                                MailboxMatcher.select(
-                                    MailboxAddress.parse(to),
-                                    from: surfaces,
-                                    identity: { $0.identity }
-                                )
-                            )
-                        ])
-                    }
-                }
-            }
-
-            let resolver = MailboxGlobalResolver(surfaces: { surfaces })
-            let resolution = resolver.resolve(
-                name: to,
-                senderWorkspaceId: senderWorkspaceId,
-                workspaceQualifier: qualifierUUID
-            )
-
-            let candidates = self.mailboxCandidatePayload(
-                MailboxMatcher.select(
-                    MailboxAddress.parse(to),
-                    from: surfaces,
-                    identity: { $0.identity }
-                )
-            )
-            switch resolution {
-            case .unique(let workspaceId, let surfaceIds):
-                return .ok([
-                    "resolution": "unique",
-                    "target_workspace_id": workspaceId.uuidString,
-                    "target_workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-                    "surface_ids": surfaceIds.map(\.uuidString),
-                    "candidates": candidates
-                ])
-            case .ambiguous:
-                return .ok([
-                    "resolution": "ambiguous",
-                    "candidates": candidates
-                ])
-            case .unresolved:
-                return .ok([
-                    "resolution": "unresolved",
-                    "candidates": candidates
-                ])
-            }
-        }
-    }
 
     private func mailboxCandidatePayload(
         _ surfaces: [MailboxGlobalResolver.Surface]
@@ -10285,293 +8712,11 @@ class TerminalController {
         return nil
     }
 
-    private func v2ConversationClaim(params: [String: Any]) -> V2CallResult {
-        if ConversationStorePolicy.isDisabled {
-            return .ok(["disabled": true, "kill_switch": "CMUX_DISABLE_CONVERSATION_STORE"])
-        }
-        guard let kind = v2String(params, "kind"), !kind.isEmpty else {
-            return .err(code: "invalid_kind", message: "kind required", data: nil)
-        }
-        let surfaceResult = v2ResolveSurfaceForConversation(params: params)
-        guard case .success(let surfaceId) = surfaceResult else {
-            if case .failure(let err) = surfaceResult { return err }
-            return .err(code: "internal_error", message: "surface resolution", data: nil)
-        }
-        let cwd = v2String(params, "cwd")
-        let placeholder = v2String(params, "placeholder_id")
-            ?? "wrapper-claim:\(surfaceId.uuidString):\(Int(Date().timeIntervalSince1970))"
 
-        let result = conversationStoreSync { store in
-            await store.claim(
-                surfaceId: surfaceId.uuidString,
-                kind: kind,
-                cwd: cwd,
-                placeholderId: placeholder
-            )
-        }
-        // Bump the surface activity timestamp; the wrapper-claim establishes
-        // the lower bound used by the Codex scrape filter.
-        SurfaceActivityTracker.shared.recordActivity(surfaceId: surfaceId.uuidString)
-        guard let ref = result else {
-            return .err(code: "internal_error", message: "store timeout", data: nil)
-        }
-        return .ok([
-            "surface_id": surfaceId.uuidString,
-            "kind": ref.kind,
-            "id": ref.id,
-            "placeholder": ref.placeholder,
-            "captured_via": ref.capturedVia.rawValue,
-            "state": ref.state.rawValue
-        ])
-    }
 
-    private func v2ConversationPush(params: [String: Any]) -> V2CallResult {
-        if ConversationStorePolicy.isDisabled {
-            return .ok(["disabled": true, "kill_switch": "CMUX_DISABLE_CONVERSATION_STORE"])
-        }
-        guard let kind = v2String(params, "kind"), !kind.isEmpty else {
-            return .err(code: "invalid_kind", message: "kind required", data: nil)
-        }
-        guard let id = v2String(params, "id"), !id.isEmpty else {
-            return .err(code: "invalid_id", message: "id required", data: nil)
-        }
-        guard let sourceStr = v2String(params, "source"),
-              let source = CaptureSource(rawValue: sourceStr) else {
-            return .err(code: "invalid_source",
-                        message: "source must be one of hook, scrape, manual, wrapperClaim",
-                        data: nil)
-        }
-        // Validate id grammar against the strategy if registered.
-        let registry = ConversationStrategyRegistry.v1
-        if let strategy = registry.strategy(forKind: kind), !strategy.isValidId(id) {
-            return .err(code: "invalid_id_grammar",
-                        message: "id does not match the \(kind) strategy's grammar",
-                        data: ["kind": kind])
-        }
-        let surfaceResult = v2ResolveSurfaceForConversation(params: params)
-        guard case .success(let surfaceId) = surfaceResult else {
-            if case .failure(let err) = surfaceResult { return err }
-            return .err(code: "internal_error", message: "surface resolution", data: nil)
-        }
-        let cwd = v2String(params, "cwd")
-        let reason = v2String(params, "diagnostic_reason")
-        let stateRaw = (v2String(params, "state") ?? "alive").lowercased()
-        // C11-24 review (I2): strict validation. Silent fallthrough to
-        // .alive swallowed typos ("susended" → .alive) and corrupted the
-        // state machine.
-        let state: ConversationState
-        switch stateRaw {
-        case "ended": state = .unknown // SessionEnd → unknown; next-launch scrape reclassifies
-        case "alive": state = .alive
-        case "suspended": state = .suspended
-        case "tombstoned": state = .tombstoned
-        case "unknown": state = .unknown
-        case "unsupported": state = .unsupported
-        default:
-            return .err(code: "invalid_state",
-                        message: "state must be one of alive, suspended, ended, tombstoned, unknown, unsupported",
-                        data: ["state": stateRaw])
-        }
-        // C11-24 review (I1): payload coercion + size cap.
-        // - Bool bridges to NSNumber on Apple platforms; the previous
-        //   ordering (NSNumber before Bool) silently coerced
-        //   `{"is_async": true}` to `.number(1.0)`. Check Bool first via
-        //   CFBoolean type id.
-        // - The else-branch is now an explicit reject so nested
-        //   objects/arrays/null don't disappear.
-        // - Cap the total serialised payload at 64 KiB to match the
-        //   metadata path; oversized hook input shouldn't be allowed to
-        //   land in the snapshot.
-        let payload: [String: PersistedJSONValue]?
-        if let dict = params["payload"] as? [String: Any] {
-            var out: [String: PersistedJSONValue] = [:]
-            out.reserveCapacity(dict.count)
-            for (k, v) in dict {
-                if let b = conversationBoolValue(v) {
-                    out[k] = .bool(b)
-                } else if let s = v as? String {
-                    out[k] = .string(s)
-                } else if let n = v as? NSNumber {
-                    out[k] = .number(n.doubleValue)
-                } else {
-                    return .err(code: "invalid_payload",
-                                message: "payload values must be string, number, or bool",
-                                data: ["key": k])
-                }
-            }
-            if let data = try? JSONSerialization.data(withJSONObject: dict),
-               data.count > Self.conversationPayloadMaxBytes {
-                return .err(code: "payload_too_large",
-                            message: "payload exceeds 64 KiB cap",
-                            data: ["bytes": data.count])
-            }
-            payload = out
-        } else if params["payload"] != nil {
-            return .err(code: "invalid_payload",
-                        message: "payload must be an object",
-                        data: nil)
-        } else {
-            payload = nil
-        }
 
-        let ref = conversationStoreSync { store in
-            await store.push(
-                surfaceId: surfaceId.uuidString,
-                kind: kind,
-                id: id,
-                source: source,
-                cwd: cwd,
-                state: state,
-                diagnosticReason: reason,
-                payload: payload
-            )
-        }
-        SurfaceActivityTracker.shared.recordActivity(surfaceId: surfaceId.uuidString)
-        guard let ref else {
-            return .err(code: "internal_error", message: "store timeout", data: nil)
-        }
-        return .ok([
-            "surface_id": surfaceId.uuidString,
-            "kind": ref.kind,
-            "id": ref.id,
-            "captured_via": ref.capturedVia.rawValue,
-            "state": ref.state.rawValue
-        ])
-    }
 
-    private func v2ConversationTombstone(params: [String: Any]) -> V2CallResult {
-        // C11-24 review (B4): the CLI requires --kind and --id and the
-        // operator typically targets a specific stale ref. Tombstoning
-        // the active ref unconditionally is a footgun: typing
-        // `c11 conversation tombstone --kind codex --id <old>` from a
-        // surface whose active ref is now a different conversation used
-        // to wipe the current one. Validate kind+id against the active
-        // ref before mutating.
-        guard let kind = v2String(params, "kind"), !kind.isEmpty else {
-            return .err(code: "invalid_kind", message: "kind required", data: nil)
-        }
-        guard let id = v2String(params, "id"), !id.isEmpty else {
-            return .err(code: "invalid_id", message: "id required", data: nil)
-        }
-        let surfaceResult = v2ResolveSurfaceForConversation(params: params)
-        guard case .success(let surfaceId) = surfaceResult else {
-            if case .failure(let err) = surfaceResult { return err }
-            return .err(code: "internal_error", message: "surface resolution", data: nil)
-        }
-        let reason = v2String(params, "reason")
-        // The bridge wraps each result in Optional (nil = store timeout),
-        // and `active(for:)` itself returns Optional. Unwrap both.
-        let activeOpt: ConversationRef?? = conversationStoreSync { store in
-            await store.active(for: surfaceId.uuidString)
-        }
-        guard case .some(let maybeActive) = activeOpt else {
-            return .err(code: "internal_error", message: "store timeout", data: nil)
-        }
-        guard let active = maybeActive else {
-            return .err(code: "not_found",
-                        message: "no active conversation for surface",
-                        data: ["surface_id": surfaceId.uuidString])
-        }
-        if active.kind != kind || active.id != id {
-            return .err(code: "id_mismatch",
-                        message: "active ref kind/id does not match request",
-                        data: [
-                            "surface_id": surfaceId.uuidString,
-                            "active_kind": active.kind,
-                            "active_id": active.id,
-                            "requested_kind": kind,
-                            "requested_id": id
-                        ])
-        }
-        _ = conversationStoreSync { store -> Void in
-            await store.tombstone(surfaceId: surfaceId.uuidString, reason: reason)
-        }
-        return .ok([
-            "surface_id": surfaceId.uuidString,
-            "kind": kind,
-            "id": id,
-            "result": "tombstoned"
-        ])
-    }
 
-    private func v2ConversationList(params: [String: Any]) -> V2CallResult {
-        let snap = conversationStoreSync { store in
-            await store.snapshot()
-        } ?? [:]
-        let surfaceFilter = v2UUID(params, "surface_id")
-        var entries: [[String: Any]] = []
-        for (sid, surfaceConv) in snap {
-            if let f = surfaceFilter, sid != f.uuidString { continue }
-            if let active = surfaceConv.active {
-                entries.append([
-                    "surface_id": sid,
-                    "kind": active.kind,
-                    "id": active.id,
-                    "placeholder": active.placeholder,
-                    "captured_via": active.capturedVia.rawValue,
-                    "state": active.state.rawValue,
-                    "captured_at": active.capturedAt.timeIntervalSince1970,
-                    "diagnostic_reason": v2OrNull(active.diagnosticReason),
-                    "cwd": v2OrNull(active.cwd)
-                ])
-            }
-        }
-        // C11-24 review (M3): expose kill-switch state so operators
-        // diagnosing "why isn't my session resuming?" can confirm
-        // whether the store path is even active.
-        return .ok([
-            "conversations": entries,
-            "is_disabled": ConversationStorePolicy.isDisabled
-        ])
-    }
-
-    private func v2ConversationGet(params: [String: Any]) -> V2CallResult {
-        let surfaceResult = v2ResolveSurfaceForConversation(params: params)
-        guard case .success(let surfaceId) = surfaceResult else {
-            if case .failure(let err) = surfaceResult { return err }
-            return .err(code: "internal_error", message: "surface resolution", data: nil)
-        }
-        let surfaceConv = conversationStoreSync { store in
-            await store.conversations(for: surfaceId.uuidString)
-        }
-        guard let surfaceConv else {
-            return .err(code: "internal_error", message: "store timeout", data: nil)
-        }
-        var out: [String: Any] = [
-            "surface_id": surfaceId.uuidString,
-            "active": NSNull(),
-            "can_resume": false,
-            "history": surfaceConv.history.map { conversationRefAsDict($0) }
-        ]
-        if let active = surfaceConv.active {
-            out["active"] = conversationRefAsDict(active)
-            let registry = ConversationStrategyRegistry.v1
-            if let strategy = registry.strategy(forKind: active.kind) {
-                if case .skip = strategy.resume(ref: active) {
-                    out["can_resume"] = false
-                } else {
-                    out["can_resume"] = true
-                }
-            }
-        }
-        return .ok(out)
-    }
-
-    private func v2ConversationClear(params: [String: Any]) -> V2CallResult {
-        let surfaceResult = v2ResolveSurfaceForConversation(params: params)
-        guard case .success(let surfaceId) = surfaceResult else {
-            if case .failure(let err) = surfaceResult { return err }
-            return .err(code: "internal_error", message: "surface resolution", data: nil)
-        }
-        _ = conversationStoreSync { store -> Void in
-            await store.clear(surfaceId: surfaceId.uuidString)
-        }
-        SurfaceActivityTracker.shared.clear(surfaceId: surfaceId.uuidString)
-        return .ok([
-            "surface_id": surfaceId.uuidString,
-            "result": "cleared"
-        ])
-    }
 
     private func conversationRefAsDict(_ ref: ConversationRef) -> [String: Any] {
         return [
@@ -10588,272 +8733,17 @@ class TerminalController {
 
     // MARK: - V2 Notification Methods
 
-    private func v2NotificationCreate(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
 
-        let title = (params["title"] as? String) ?? "Notification"
-        let subtitle = (params["subtitle"] as? String) ?? ""
-        let body = (params["body"] as? String) ?? ""
 
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            let surfaceId = ws.focusedPanelId
-            TerminalNotificationStore.shared.addNotification(
-                tabId: ws.id,
-                surfaceId: surfaceId,
-                title: title,
-                subtitle: subtitle,
-                body: body
-            )
-            result = .ok(["workspace_id": ws.id.uuidString, "surface_id": v2OrNull(surfaceId?.uuidString)])
-        }
-        return result
-    }
 
-    private func v2NotificationCreateForSurface(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let surfaceId = v2UUID(params, "surface_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid surface_id", data: nil)
-        }
 
-        let title = (params["title"] as? String) ?? "Notification"
-        let subtitle = (params["subtitle"] as? String) ?? ""
-        let body = (params["body"] as? String) ?? ""
 
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            guard ws.panels[surfaceId] != nil else {
-                result = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
-                return
-            }
-            TerminalNotificationStore.shared.addNotification(
-                tabId: ws.id,
-                surfaceId: surfaceId,
-                title: title,
-                subtitle: subtitle,
-                body: body
-            )
-            result = .ok(["workspace_id": ws.id.uuidString, "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id), "surface_id": surfaceId.uuidString, "surface_ref": v2Ref(kind: .surface, uuid: surfaceId), "window_id": v2OrNull(v2ResolveWindowId(tabManager: tabManager)?.uuidString), "window_ref": v2Ref(kind: .window, uuid: v2ResolveWindowId(tabManager: tabManager))])
-        }
-        return result
-    }
 
-    private func v2NotificationCreateForTarget(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let wsId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        guard let surfaceId = v2UUID(params, "surface_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid surface_id", data: nil)
-        }
 
-        let title = (params["title"] as? String) ?? "Notification"
-        let subtitle = (params["subtitle"] as? String) ?? ""
-        let body = (params["body"] as? String) ?? ""
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to notify", data: nil)
-        v2MainSync {
-            guard let ws = tabManager.tabs.first(where: { $0.id == wsId }) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: ["workspace_id": wsId.uuidString])
-                return
-            }
-            guard ws.panels[surfaceId] != nil else {
-                result = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
-                return
-            }
-            TerminalNotificationStore.shared.addNotification(
-                tabId: ws.id,
-                surfaceId: surfaceId,
-                title: title,
-                subtitle: subtitle,
-                body: body
-            )
-            result = .ok(["workspace_id": ws.id.uuidString, "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id), "surface_id": surfaceId.uuidString, "surface_ref": v2Ref(kind: .surface, uuid: surfaceId), "window_id": v2OrNull(v2ResolveWindowId(tabManager: tabManager)?.uuidString), "window_ref": v2Ref(kind: .window, uuid: v2ResolveWindowId(tabManager: tabManager))])
-        }
-        return result
-    }
-
-    private func v2NotificationList() -> [String: Any] {
-        var items: [[String: Any]] = []
-        v2MainSync {
-            items = TerminalNotificationStore.shared.notifications.map { n in
-                return [
-                    "id": n.id.uuidString,
-                    "workspace_id": n.tabId.uuidString,
-                    "surface_id": v2OrNull(n.surfaceId?.uuidString),
-                    "is_read": n.isRead,
-                    "title": n.title,
-                    "subtitle": n.subtitle,
-                    "body": n.body
-                ]
-            }
-        }
-        return ["notifications": items]
-    }
-
-    private func v2NotificationClear() -> V2CallResult {
-        v2MainSync {
-            TerminalNotificationStore.shared.clearAll()
-        }
-        return .ok([:])
-    }
-
-    private func v2FeedbackOpen(params: [String: Any]) -> V2CallResult {
-        let workspaceId = v2UUID(params, "workspace_id")
-        let windowId = v2UUID(params, "window_id")
-        let shouldActivate = v2Bool(params, "activate") ?? false
-        DispatchQueue.main.async {
-            let targetWindow: NSWindow?
-            if let windowId, let app = AppDelegate.shared {
-                targetWindow = app.mainWindow(for: windowId)
-            } else if let workspaceId, let app = AppDelegate.shared {
-                targetWindow = app.mainWindowContainingWorkspace(workspaceId)
-            } else {
-                targetWindow = nil
-            }
-
-            if shouldActivate {
-                if let targetWindow {
-                    targetWindow.makeKeyAndOrderFront(nil)
-                    NSRunningApplication.current.activate(options: [.activateAllWindows, .activateIgnoringOtherApps])
-                } else {
-                    NSRunningApplication.current.activate(options: [.activateAllWindows, .activateIgnoringOtherApps])
-                }
-            }
-
-            FeedbackComposerBridge.openComposer(in: targetWindow)
-        }
-        return .ok(["opened": true])
-    }
-
-    private func v2SettingsOpen(params: [String: Any]) -> V2CallResult {
-        let targetRaw = v2String(params, "target")
-        let shouldActivate = v2Bool(params, "activate") ?? true
-
-        let navigationTarget: SettingsNavigationTarget?
-        switch targetRaw {
-        case nil:
-            navigationTarget = nil
-        case SettingsNavigationTarget.keyboardShortcuts.rawValue:
-            navigationTarget = .keyboardShortcuts
-        default:
-            return .err(code: "invalid_params", message: "Unknown settings target", data: ["target": targetRaw ?? ""])
-        }
-
-        DispatchQueue.main.async {
-            if shouldActivate {
-                AppDelegate.presentPreferencesWindow(navigationTarget: navigationTarget)
-            } else {
-                SettingsWindowController.shared.show(navigationTarget: navigationTarget)
-            }
-        }
-        return .ok([
-            "opened": true,
-            "target": navigationTarget?.rawValue ?? "general",
-        ])
-    }
-
-    private func v2FeedbackSubmit(params: [String: Any]) -> V2CallResult {
-        guard let email = params["email"] as? String else {
-            return .err(code: "invalid_params", message: "Missing email", data: ["field": "email"])
-        }
-        guard let body = params["body"] as? String else {
-            return .err(code: "invalid_params", message: "Missing body", data: ["field": "body"])
-        }
-        let imagePaths = params["image_paths"] as? [String] ?? []
-
-        let semaphore = DispatchSemaphore(value: 0)
-        var result: V2CallResult = .err(code: "internal_error", message: "Feedback submission failed", data: nil)
-
-        Task {
-            let resolved: V2CallResult
-            do {
-                let attachmentCount = try await FeedbackComposerBridge.submit(
-                    email: email,
-                    message: body,
-                    imagePaths: imagePaths
-                )
-                resolved = .ok([
-                    "submitted": true,
-                    "attachment_count": attachmentCount,
-                ])
-            } catch let error as FeedbackComposerBridgeError {
-                let code: String
-                switch error {
-                case .invalidEmail, .emptyMessage, .messageTooLong, .tooManyImages, .invalidImagePath:
-                    code = "invalid_params"
-                case .submissionFailed:
-                    code = "request_failed"
-                }
-                resolved = .err(code: code, message: error.localizedDescription, data: nil)
-            } catch {
-                resolved = .err(code: "internal_error", message: error.localizedDescription, data: nil)
-            }
-
-            result = resolved
-            semaphore.signal()
-        }
-
-        if semaphore.wait(timeout: .now() + 35) == .timedOut {
-            return .err(code: "timeout", message: "Feedback submission timed out", data: nil)
-        }
-
-        return result
-    }
 
     // MARK: - V2 App Focus Methods
 
-    private func v2AppFocusOverride(params: [String: Any]) -> V2CallResult {
-        // Accept either:
-        // - state: "active" | "inactive" | "clear"
-        // - focused: true/false/null
-        if let state = v2String(params, "state")?.lowercased() {
-            switch state {
-            case "active":
-                AppFocusState.overrideIsFocused = true
-            case "inactive":
-                AppFocusState.overrideIsFocused = false
-            case "clear", "none":
-                AppFocusState.overrideIsFocused = nil
-            default:
-                return .err(code: "invalid_params", message: "Invalid state (active|inactive|clear)", data: ["state": state])
-            }
-        } else if params.keys.contains("focused") {
-            if let focused = v2Bool(params, "focused") {
-                AppFocusState.overrideIsFocused = focused
-            } else {
-                AppFocusState.overrideIsFocused = nil
-            }
-        } else {
-            return .err(code: "invalid_params", message: "Missing state or focused", data: nil)
-        }
 
-        let overrideVal: Any = v2OrNull(AppFocusState.overrideIsFocused.map { $0 as Any })
-        return .ok(["override": overrideVal])
-    }
-
-    private func v2AppSimulateActive() -> V2CallResult {
-        v2MainSync {
-            AppDelegate.shared?.applicationDidBecomeActive(
-                Notification(name: NSApplication.didBecomeActiveNotification)
-            )
-        }
-        return .ok([:])
-    }
 
     // MARK: - V2 Browser Methods
 
@@ -11379,182 +9269,9 @@ class TerminalController {
 
     // MARK: - Markdown
 
-    private func v2MarkdownOpen(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let rawPath = v2String(params, "path") else {
-            return .err(code: "invalid_params", message: "Missing 'path' parameter", data: nil)
-        }
-
-        // Resolve the path (expand ~ and standardize)
-        let expandedPath = NSString(string: rawPath).expandingTildeInPath
-        let filePath = NSString(string: expandedPath).standardizingPath
-
-        // Reject paths that aren't absolute after resolution
-        guard filePath.hasPrefix("/") else {
-            return .err(code: "invalid_params", message: "Path must be absolute: \(filePath)", data: ["path": filePath])
-        }
-
-        // Validate the file exists and is a regular file (not a directory)
-        var isDir: ObjCBool = false
-        guard FileManager.default.fileExists(atPath: filePath, isDirectory: &isDir) else {
-            return .err(code: "not_found", message: "File not found: \(filePath)", data: ["path": filePath])
-        }
-        guard !isDir.boolValue else {
-            return .err(code: "invalid_params", message: "Path is a directory, not a file: \(filePath)", data: ["path": filePath])
-        }
-        guard FileManager.default.isReadableFile(atPath: filePath) else {
-            return .err(code: "permission_denied", message: "File not readable: \(filePath)", data: ["path": filePath])
-        }
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to create markdown panel", data: nil)
-        v2MainSync {
-            // M6 — if pane_id is supplied, locate the owning workspace across all
-            // windows so `markdown.open --pane P` works standalone (spec: --pane
-            // uniquely identifies). This takes precedence over workspace_id/window_id
-            // (which may be injected from env vars by the CLI).
-            var resolvedTabManager: TabManager = tabManager
-            var resolvedWorkspace: Workspace?
-            if v2HasNonNullParam(params, "pane_id") {
-                guard let paneUUID = v2UUID(params, "pane_id") else {
-                    result = .err(code: "invalid_params", message: "Invalid pane_id", data: nil)
-                    return
-                }
-                if let located = AppDelegate.shared?.locatePane(paneId: paneUUID) {
-                    resolvedWorkspace = located.workspace
-                    resolvedTabManager = located.tabManager
-                }
-            }
-            guard let ws = resolvedWorkspace ?? v2ResolveWorkspace(params: params, tabManager: resolvedTabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            v2MaybeFocusWindow(for: resolvedTabManager)
-            v2MaybeSelectWorkspace(resolvedTabManager, workspace: ws)
-
-            // M6 — if pane_id is supplied, open as a tab inside that pane (no split).
-            if v2HasNonNullParam(params, "pane_id") {
-                guard let paneUUID = v2UUID(params, "pane_id") else {
-                    result = .err(code: "invalid_params", message: "Invalid pane_id", data: nil)
-                    return
-                }
-                guard let targetPaneId = ws.bonsplitController.allPaneIds.first(where: { $0.id == paneUUID }) else {
-                    result = .err(code: "not_found", message: "Pane not found in workspace", data: ["pane_id": paneUUID.uuidString])
-                    return
-                }
-
-                let createdPanel = ws.newMarkdownSurface(
-                    inPane: targetPaneId,
-                    filePath: filePath,
-                    focus: v2FocusAllowed()
-                )
-
-                guard let markdownPanelId = createdPanel?.id else {
-                    result = .err(code: "internal_error", message: "Failed to create markdown panel", data: nil)
-                    return
-                }
-
-                let windowId = v2ResolveWindowId(tabManager: resolvedTabManager)
-                result = .ok([
-                    "window_id": v2OrNull(windowId?.uuidString),
-                    "window_ref": v2Ref(kind: .window, uuid: windowId),
-                    "workspace_id": ws.id.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                    "pane_id": targetPaneId.id.uuidString,
-                    "pane_ref": v2Ref(kind: .pane, uuid: targetPaneId.id),
-                    "surface_id": markdownPanelId.uuidString,
-                    "surface_ref": v2Ref(kind: .surface, uuid: markdownPanelId),
-                    "target_pane_id": targetPaneId.id.uuidString,
-                    "target_pane_ref": v2Ref(kind: .pane, uuid: targetPaneId.id),
-                    "path": filePath
-                ])
-                return
-            }
-
-            let sourceSurfaceId = v2UUID(params, "surface_id") ?? ws.focusedPanelId
-            guard let sourceSurfaceId else {
-                result = .err(code: "not_found", message: "No focused surface to split", data: nil)
-                return
-            }
-            guard ws.panels[sourceSurfaceId] != nil else {
-                result = .err(code: "not_found", message: "Source surface not found", data: ["surface_id": sourceSurfaceId.uuidString])
-                return
-            }
-
-            let sourcePaneUUID = ws.paneId(forPanelId: sourceSurfaceId)?.id
-
-            let createdPanel = ws.newMarkdownSplit(
-                from: sourceSurfaceId,
-                orientation: .horizontal,
-                filePath: filePath,
-                focus: v2FocusAllowed()
-            )
-
-            guard let markdownPanelId = createdPanel?.id else {
-                result = .err(code: "internal_error", message: "Failed to create markdown panel", data: nil)
-                return
-            }
-
-            let targetPaneUUID = ws.paneId(forPanelId: markdownPanelId)?.id
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "pane_id": v2OrNull(targetPaneUUID?.uuidString),
-                "pane_ref": v2Ref(kind: .pane, uuid: targetPaneUUID),
-                "surface_id": markdownPanelId.uuidString,
-                "surface_ref": v2Ref(kind: .surface, uuid: markdownPanelId),
-                "source_surface_id": sourceSurfaceId.uuidString,
-                "source_surface_ref": v2Ref(kind: .surface, uuid: sourceSurfaceId),
-                "source_pane_id": v2OrNull(sourcePaneUUID?.uuidString),
-                "source_pane_ref": v2Ref(kind: .pane, uuid: sourcePaneUUID),
-                "target_pane_id": v2OrNull(targetPaneUUID?.uuidString),
-                "target_pane_ref": v2Ref(kind: .pane, uuid: targetPaneUUID),
-                "path": filePath
-            ])
-        }
-        return result
-    }
 
     // MARK: - M3 — structured sidebar.state
 
-    /// Structured `sidebar_state` — JSON variant that includes the M3 `agent_chip` block.
-    private func v2SidebarState(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        var payload: [String: Any]?
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else { return }
-            let statusEntries = ws.sidebarStatusEntriesInDisplayOrder()
-            let metadataBlocks = ws.sidebarMetadataBlocksInDisplayOrder()
-
-            var out: [String: Any] = [
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "status_count": statusEntries.count,
-                "meta_block_count": metadataBlocks.count,
-                "log_count": ws.logEntries.count
-            ]
-            if let focused = ws.focusedPanelId {
-                out["focused_surface_id"] = focused.uuidString
-                out["focused_surface_ref"] = v2Ref(kind: .surface, uuid: focused)
-            }
-
-            // Agent chip resolution.
-            let chipDict: [String: Any] = self.resolveAgentChipDict(workspace: ws)
-            out["agent_chip"] = chipDict
-
-            payload = out
-        }
-        guard let out = payload else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-        return .ok(out)
-    }
 
     /// Build the agent_chip payload for inclusion in `sidebar.state` (v2) and `sidebar_state` (v1 text).
     private func resolveAgentChipDict(workspace ws: Workspace) -> [String: Any] {
@@ -11631,52 +9348,6 @@ class TerminalController {
 
     // MARK: - M6 — markdown.get_content
 
-    private func v2MarkdownGetContent(params: [String: Any]) -> V2CallResult {
-        guard let resolved = v2ResolveWorkspaceSurface(params: params) else {
-            return .err(code: "not_found", message: "Surface not found", data: nil)
-        }
-        let (ws, surfaceId) = resolved
-
-        var payload: [String: Any]?
-        var errResult: V2CallResult?
-        v2MainSync {
-            guard let panel = ws.panels[surfaceId] else {
-                errResult = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
-                return
-            }
-            guard let markdown = panel as? MarkdownPanel else {
-                errResult = .err(code: "invalid_params", message: "Surface is not a markdown panel", data: ["surface_id": surfaceId.uuidString])
-                return
-            }
-
-            let content = markdown.content
-            let contentBytes = content.data(using: .utf8) ?? Data()
-            let sha = SHA256.hash(data: contentBytes).map { String(format: "%02x", $0) }.joined()
-            let softCap = 256 * 1024
-
-            var out: [String: Any] = [
-                "surface_id": surfaceId.uuidString,
-                "surface_ref": v2Ref(kind: .surface, uuid: surfaceId),
-                "type": PanelType.markdown.rawValue,
-                "file_path": markdown.filePath,
-                "content_length": contentBytes.count,
-                "content_sha256": sha,
-                "is_file_unavailable": markdown.isFileUnavailable
-            ]
-            if contentBytes.count > softCap {
-                out["truncated"] = true
-                out["reason"] = "content_too_large"
-            } else {
-                out["content"] = content
-            }
-            payload = out
-        }
-        if let errResult { return errResult }
-        guard let out = payload else {
-            return .err(code: "not_found", message: "Surface not found", data: nil)
-        }
-        return .ok(out)
-    }
 
     // MARK: - Browser
